Added files via upload
a1.c:
l implemented a process monitoring system (PMS), interacting with the operating system using system calls. The PMS includes two parts: a process generator and a process monitor. The former is to randomly generate a set of processes running in the background; the latter is to monitor, suspend, resume, or terminate these processes.