Added files via upload
Part 1 diskinfo:
It displays information about the file system. It involves the understanding of the file system structure of MS-DOS, including FAT Partition Boot Sector, FAT File Allocation Table, FAT Root Folder, FAT Folder Structure, and so on.

Part 2 disklist:
It displays the contents of the root directory in the file system.

Part 3 diskget:
It copies a file from the file system to the current directory in Linux. If the specified file is not found in the root directory of the file system, you should output the message File not found and exit.

Part 4 diskput:
It copies a file from the current Linux directory into the root directory of the file system. If the specified file is not found, you should output the message File not found. on a single line and exit. If the file system does not have enough free space to store the file, you should output the message No enough free space in the disk image and exit.